_parsing: Lift the parsing of remaining length into parse.
Every MQTT packet has a fixed header, at least one byte of which
represents the remaining length of the packet.  There is no value, and
likely real cost in having this embedded in every parser, so pull it
to the top level.

Additionally, implement the multibyte encoding parsing, though without
tests at the moment.