Elimintate MQTTPacket, let each parser return their own.
Rather than create two objects, simply to have "uniformity", ensure
each returned class has a pkt_type attribute and the fields relevant
to it specifically.  This does require consumers of the parsed packet
know something about the packet, but that is likely to be true once
the pkt_type in inspected.

An alternative would be an interface, but at the moment I am trying to
keep dependencies light.